Allow specifying existing GIDs
Before, using an existing GID would result in an error when creating the
user.

When the entry point script creates a new group, it uses the GID as id
and name. adduser expects the group name. If the group already existed before,
then it most likely didn't have the id as its name, so adduser fails.
Then in turn executing s3fs fails because the user is not found.

This patch detects if the group is already present and uses it in that case.